requirements.txt
# Discord Bot Dependencies
discord.py>=2.3.0

# HTTP Requests and WebSocket
requests>=2.31.0
websocket-client>=1.6.0
aiohttp>=3.8.0

# Image Processing
Pillow>=9.0.0
opencv-python>=4.8.0

# Video Processing (for future video features)
moviepy>=1.0.3

# Configuration and Environment
python-dotenv>=1.0.0
pydantic>=2.0.0

# Async Support
asyncio-mqtt>=0.13.0

# Logging and Utilities
colorlog>=6.7.0
python-dateutil>=2.8.0

# File Management
pathlib2>=2.3.7

# JSON Schema Validation
jsonschema>=4.19.0

# Type Hints Support
typing-extensions>=4.7.0

# Real-time WebSocket communication with ComfyUI
websockets>=11.0.0

# Testing (v2.0)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Documentation (v2.0)
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0